+++ src/core/lombok/eclipse/handlers/HandleBuilder.java
@@ -145,7 +145,10 @@ public class HandleBuilder extends EclipseAnnotationHandler<Builder> {
 		List<BuilderFieldData> builderFields;
 		AccessLevel accessInners, accessOuters;
 		boolean oldFluent, oldChain, toBuilder;
-		
+
+		String toBuilderMethodName;
+		char[] toBuilderMethodNameArr;
+
 		EclipseNode builderType;
 		String builderClassName;
 		char[] builderClassNameArr;
@@ -153,6 +156,11 @@ public class HandleBuilder extends EclipseAnnotationHandler<Builder> {
 		void setBuilderClassName(String builderClassName) {
 			this.builderClassName = builderClassName;
 			this.builderClassNameArr = builderClassName.toCharArray();
+		}
+
+		void setToBuilderMethodName(String toBuilderMethodName) {
+			this.toBuilderMethodName = toBuilderMethodName;
+			this.toBuilderMethodNameArr = toBuilderMethodName.toCharArray();
 		}
 		
 		TypeParameter[] copyTypeParams() {
@@ -196,7 +204,8 @@ public class HandleBuilder extends EclipseAnnotationHandler<Builder> {
 			buildMethodName = ann.buildMethodName();
 			setBuilderClassName(getBuilderClassNameTemplate(node, ann.builderClassName()));
 			toBuilder = ann.toBuilder();
-			
+			setToBuilderMethodName(getToBuilderMethodName(node, ann.toBuilderMethodName()));
+
 			if (builderMethodName == null) builderMethodName = "builder";
 			if (buildMethodName == null) buildMethodName = "build";
 		}
@@ -206,6 +215,13 @@ public class HandleBuilder extends EclipseAnnotationHandler<Builder> {
 			override = node.getAst().readConfiguration(ConfigurationKeys.BUILDER_CLASS_NAME);
 			if (override != null && !override.isEmpty()) return override;
 			return "*Builder";
+		}
+
+		static String getToBuilderMethodName(EclipseNode node, String override) {
+			if (override != null && !override.isEmpty()) return override;
+			override = node.getAst().readConfiguration(ConfigurationKeys.TO_BUILDER_METHOD_NAME);
+			if (override != null && !override.isEmpty()) return override;
+			return TO_BUILDER_METHOD_NAME_STRING;
 		}
 		
 		MethodDeclaration createNewMethodDeclaration() {
@@ -608,9 +624,9 @@ public class HandleBuilder extends EclipseAnnotationHandler<Builder> {
 			if (md != null) injectMethod(job.parentType, md);
 		}
 		
-		if (job.toBuilder) switch (methodExists(TO_BUILDER_METHOD_NAME_STRING, job.parentType, 0)) {
+		if (job.toBuilder) switch (methodExists(job.toBuilderMethodName, job.parentType, 0)) {
 		case EXISTS_BY_USER:
-			annotationNode.addWarning("Not generating toBuilder() as it already exists.");
+			annotationNode.addWarning("Not generating " + job.toBuilderMethodName + "() as it already exists.");
 			break;
 		case NOT_EXISTS:
 			TypeParameter[] tps = job.typeParams;
@@ -667,7 +683,7 @@ public class HandleBuilder extends EclipseAnnotationHandler<Builder> {
 		long p = job.getPos();
 		
 		MethodDeclaration out = job.createNewMethodDeclaration();
-		out.selector = TO_BUILDER_METHOD_NAME;
+		out.selector = job.toBuilderMethodNameArr;
 		out.modifiers = toEclipseModifier(job.accessOuters);
 		out.bits |= ECLIPSE_DO_NOT_TOUCH_FLAG;